requirements.txt
discord.py
pymongo
python-dotenv
black
requests
